+++ src/Options/DecodingOption.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace MAChitgarha\Json\Options;
+
+/**
+ * Option provider for Json::decode().
+ *
+ * @see Json::decode()
+ */
+class DecodingOption
+{
+}